refactor(ops): add unified attention facade
